# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atiende Rentas Vacacionales — API
  version: 0.3.0
  description: |
    Contrato de API de `apps/api` (Lote 3: auth/roles/RLS/auditoría sobre el
    dominio de calendario de Lote 1). Fuente de verdad ejecutable: los
    esquemas `zod` de `apps/api/src/contrato/tipos.ts` (importables como
    `@atiende-rv/api/contrato`) — este archivo es la descripción
    human-readable del mismo contrato, publicada en el primer commit de
    este lote para que Lote 4 (frontend) pueda empezar contra tipos
    estables antes de que todos los endpoints tengan implementación
    completa.

    Reglas de dominio no negociables reflejadas en este contrato (D-006,
    D-011, BLUEPRINT): la API NUNCA expone cancelación ni modificación de
    una reserva cuyo origen sea un canal externo (Airbnb/Vrbo/Booking) —
    solo reservas DIRECTAS (creadas por esta misma API). Tampoco expone
    envío de mensajes a huéspedes (eso es Lote 6, con aprobación humana
    obligatoria). `POST /canales/:id/sync` solo ENCOLA una solicitud de
    sincronización manual en `outbox_evento`; la ejecución real la hace el
    worker de Lote 2.
servers:
  - url: http://localhost:8787
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /tenants:
    post:
      tags: [tenants]
      summary: Alta de un tenant (empresa gestora) — exclusivo de superadmin
      responses:
        "201": { description: Tenant creado }
        "403": { $ref: "#/components/responses/RolForbidden" }
...
  /propiedades:
    get:
      tags: [propiedades]
      summary: Listar propiedades visibles para el usuario autenticado (RLS)
      responses:
        "200": { description: Lista de propiedades }
    post:
      tags: [propiedades]
      summary: Crear propiedad — admin_gestora/superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearPropiedad" }
      responses:
        "201": { description: Propiedad creada }
        "403": { $ref: "#/components/responses/RolForbidden" }
        "422": { $ref: "#/components/responses/Validacion" }
  /propiedades/{id}:
    get:
      tags: [propiedades]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Propiedad }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
  /backoffice/tenants:
    get:
      tags: [backoffice]
      summary: Directorio de tenants con métricas agregadas — superadmin, SIN concesión "romper cristal"
      responses:
        "200":
          description: Directorio de tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items: { $ref: "#/components/schemas/TenantDirectorio" }
        "403": { $ref: "#/components/responses/RolForbidden" }
    post:
      tags: [backoffice]
      summary: Alta de tenant (empresa gestora) — superadmin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearTenantBackoffice" }
      responses:
        "201": { description: Tenant creado }
        "403": { $ref: "#/components/responses/RolForbidden" }
  /backoffice/tenants/{id}/suspender:
    post:
      tags: [backoffice]
      summary: Suspender un tenant con motivo — superadmin
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoSuspenderTenant" }
      responses:
        "200": { description: Tenant suspendido }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /backoffice/tenants/{id}/activar:
    post:
      tags: [backoffice]
      summary: Reactivar un tenant suspendido — superadmin
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Tenant activo }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /backoffice/romper-cristal:
    get:
      tags: [backoffice]
      summary: Concesiones "romper cristal" propias vigentes — superadmin (para el banner persistente)
      responses:
        "200":
          description: Concesiones vigentes
          content:
            application/json:
              schema:
                type: object
                properties:
                  accesos:
                    type: array
                    items: { $ref: "#/components/schemas/AccesoRomperCristal" }
    post:
      tags: [backoffice]
      summary: Crear una concesión "romper cristal" (motivo obligatorio, ventana acotada) — superadmin
      description: |
        Genera una entrada `ACCESO_ROMPER_CRISTAL` en `auditoria_mutacion`
        con actor, tenant y motivo (§Auditoría-1). Sin esta concesión,
        `is_tenant_member` deniega a un superadmin el acceso a datos de
        negocio de cualquier tenant (packages/db migración 0061).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearAccesoRomperCristal" }
      responses:
        "201": { description: Concesión creada }
        "403": { $ref: "#/components/responses/RolForbidden" }
  /backoffice/romper-cristal/{id}/revocar:
    post:
      tags: [backoffice]
      summary: Revocar una concesión propia antes de su expiración — superadmin
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Concesión revocada }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
  /backoffice/propiedades:
    get:
      tags: [backoffice]
      summary: CRUD de administración de propiedades — zona horaria/dirección/moneda
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Propiedades del tenant
          content:
            application/json:
              schema:
                type: object
                properties:
                  propiedades:
                    type: array
                    items: { $ref: "#/components/schemas/PropiedadBackoffice" }
        "403": { $ref: "#/components/responses/RolForbidden" }
    post:
      tags: [backoffice]
      summary: Alta de propiedad — zona horaria IANA, dirección mínima y moneda obligatorias (H-011)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearPropiedadBackoffice" }
      responses:
        "201": { description: Propiedad creada }
        "422": { $ref: "#/components/responses/Validacion" }
  /backoffice/propiedades/{id}:
    patch:
      tags: [backoffice]
      summary: Actualizar propiedad
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoActualizarPropiedadBackoffice" }
      responses:
        "200": { description: Propiedad actualizada }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Validacion:
      description: Error de validación de entrada (zod)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
...
    RolForbidden:
      description: "rol_forbidden: el rol/nivel de colaborador no tiene permiso para esta acción"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
    NoEncontrado:
      description: "recurso_no_encontrado (incluye recursos de otro tenant, ocultos por RLS — nunca confirma su existencia)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
  schemas:
    ErrorHttp:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum:
                - credenciales_invalidas
                - token_invalido
                - token_expirado
                - tenant_forbidden
                - rol_forbidden
                - rango_invalido
                - unidad_no_disponible
                - conflicto_pendiente
                - reserva_no_directa
                - recurso_no_encontrado
                - validacion
                - rate_limited
                - error_interno
            mensaje: { type: string }
            detalles: {}
...
    CuerpoCrearPropiedad:
      type: object
      required: [nombre, zonaHoraria]
      properties:
        nombre: { type: string }
        zonaHoraria: { type: string, description: "Zona horaria IANA, p. ej. America/Cancun" }
...
    MetricasTenant:
      type: object
      properties:
        unidadesTotal: { type: integer, minimum: 0 }
        cuentasCanalTotal: { type: integer, minimum: 0 }
        cuentasCanalConfiguradas: { type: integer, minimum: 0 }
        cuentasCanalSimulador: { type: integer, minimum: 0 }
        alertasAbiertas: { type: integer, minimum: 0 }
        outboxPendiente: { type: integer, minimum: 0 }
    TenantDirectorio:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        tipo: { type: string }
        estado: { type: string, enum: [activo, suspendido] }
        suspendidoMotivo: { type: string, nullable: true }
        creadoEn: { type: string, format: date-time }
        metricas: { $ref: "#/components/schemas/MetricasTenant" }
    CuerpoCrearTenantBackoffice:
      type: object
      required: [nombre, razonSocial]
      properties:
        nombre: { type: string }
        razonSocial: { type: string }
    CuerpoSuspenderTenant:
      type: object
      required: [motivo]
      properties:
        motivo: { type: string, minLength: 1 }
    CuerpoCrearAccesoRomperCristal:
      type: object
      required: [tenantId, motivo]
      properties:
        tenantId: { type: string, format: uuid }
        motivo: { type: string, minLength: 1 }
        alcance: { type: string, default: general }
        minutos: { type: integer, minimum: 1, maximum: 1440, default: 30 }
    AccesoRomperCristal:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        motivo: { type: string }
        alcance: { type: string }
        creadoEn: { type: string, format: date-time }
        expiraEn: { type: string, format: date-time }
        revocadoEn: { type: string, format: date-time, nullable: true }
...
    DireccionMinima:
      type: object
      required: [linea1, ciudad, pais]
      properties:
        linea1: { type: string, minLength: 1 }
        ciudad: { type: string, minLength: 1 }
        pais: { type: string, minLength: 2, maxLength: 2, description: "ISO 3166-1 alfa-2, ej. MX" }
    CuerpoCrearPropiedadBackoffice:
      type: object
      required: [nombre, zonaHoraria, moneda, direccion]
      properties:
        tenantId: { type: string, format: uuid, description: "Requerido si el actor es superadmin" }
        nombre: { type: string }
        zonaHoraria: { type: string, description: "IANA, ej. America/Mexico_City" }
        moneda: { type: string, pattern: "^[A-Z]{3}$" }
        direccion: { $ref: "#/components/schemas/DireccionMinima" }
    CuerpoActualizarPropiedadBackoffice:
      type: object
      properties:
        nombre: { type: string }
        zonaHoraria: { type: string }
        moneda: { type: string, pattern: "^[A-Z]{3}$" }
        direccion: { $ref: "#/components/schemas/DireccionMinima" }
    PropiedadBackoffice:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        nombre: { type: string }
        zonaHoraria: { type: string }
        moneda: { type: string, nullable: true }
        direccion:
          type: object
          nullable: true
          properties:
            linea1: { type: string, nullable: true }
            ciudad: { type: string, nullable: true }
            pais: { type: string, nullable: true }
...